requirements2.txt
appdirs==1.4.0
appnope==0.1.0
apsw==3.15.2.post1
argh==0.26.2
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
bleach==1.5.0
blinker==1.4
bzr==2.7.0
certifi==2015.11.20.1
cffi==1.9.1
chardet==2.3.0
CherryPy==8.8.0
click==6.7
ConfigArgParse==0.11.0
construct==2.8.10
cryptography==1.7.1
cssselect==1.0.1
cssutils==1.0.1
cycler==0.10.0
Cython==0.25.2
daft==0.0.4
decorator==4.0.11
deepin-menu==0.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10.4
django-floppyforms==1.7.0
dnspython==1.15.0
enum34==1.1.6
Flask==0.12
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.16.0
GDAL==2.1.1
gnureadline==6.3.3
gunicorn==19.6.0
h2==2.5.1
h5py==2.6.0
hpack==2.3.0
html2text==2016.9.19
html5lib==0.9999999
httplib2==0.9.2
hyperframe==4.0.1
idna==2.2
ipaddress==1.0.18
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==4.1.1
itsdangerous==0.24
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==4.0.3
jupyter-core==4.2.1
Keras==1.2.0
louis==3.0.0
lxml==3.7.2
Markdown==2.6.7
MarkupSafe==0.23
matplotlib==1.5.3
mechanize==0.2.5
mistune==0.7.1
mock==2.0.0
mutagen==1.36
nbconvert==4.1.0
nbformat==4.0.1
netifaces==0.10.5
networkx==1.11
notebook==4.0.6
numpy==1.11.3
oauth2==1.9.0.post1
oauthlib==2.0.1
olefile==0.43
packaging==16.8
pandas==0.18.1
passlib==1.7.0.post20161227062707
path.py==9.0
pathlib==1.0.1
pathtools==0.1.2
pbr==1.10.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==3.4.2
ply==3.9
prompt-toolkit==1.0.9
protobuf==3.0.0
psutil==5.0.1
psycopg2==2.6.2
ptyprocess==0.5.1
pwquality==1.3.0
pyasn1==0.1.9
pycparser==2.17
pyflakes==1.4.0
Pygments==2.1.3
PyOpenGL==3.1.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
pyperclip==1.5.27
python-dateutil==2.6.0
python-decouple==3.0
python-linkedin==4.1
python-qt==0.50
python-sane==2.8.3
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.1.1
requests==2.12.4
requests-oauthlib==0.7.0
scipy==0.18.1
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
spyder==2.3.9
static3==0.7.0
team==1.0
tensorflow==0.11.0rc2
terminado==0.5
tflearn==0.2.1
Theano==0.8.2
tornado==4.4.2
traitlets==4.3.1
Unipath==1.1
urllib3==1.19.1
urwid==1.3.1
virtualenv==15.0.3
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.11.15
whitenoise==3.2.2
xcffib==0.4.2
xpybutil==0.0.1
